Include structure and namespace changes
- Include files are now nested within a syscommon/ directory
- Namespace name changed from SysCommon to syscommon
- Renamed Utils to StringUtils and moved to the util namespace
