+++ codebase/src/cpp/syscommon/include/syscommon/net/InetSocketAddress.h
@@ -15,9 +15,9 @@
  * Portions Copyright [yyyy] [name of copyright owner]
  */
 
-#include "Platform.h"
+#include "syscommon/Platform.h"
 
-namespace SysCommon
+namespace syscommon
 {
 	/**
 	 * This class represents an IP Socket Address (IP address + port number)
