+++ codebase/src/cpp/syscommon/src/Event.cpp
@@ -12,10 +12,11 @@
  * enclosed by brackets "[]" replaced with your own identifying information:
  * Portions Copyright [yyyy] [name of copyright owner]
  */
-#include "concurrent/Event.h"
+#include "syscommon/concurrent/Event.h"
+
 #include <limits.h>
 
-using namespace SysCommon;
+using namespace syscommon;
 
 //----------------------------------------------------------
 //                      CONSTRUCTORS
