+++ codebase/src/cpp/syscommon/src/InetSocketAddress.cpp
@@ -12,9 +12,9 @@
  * enclosed by brackets "[]" replaced with your own identifying information: 
  * Portions Copyright [yyyy] [name of copyright owner]
  */
-#include "net/InetSocketAddress.h"
+#include "syscommon/net/InetSocketAddress.h"
 
-using namespace SysCommon;
+using namespace syscommon;
 
 //----------------------------------------------------------
 //                      CONSTRUCTORS
